+++ vue/src/apps/cortex/VHeader.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script setup>
- import VBrand from '@/components/header/VBrand.vue';
- import VHelpAndLogout from '@/components/header/VHelpAndLogout.vue';
- import VSimNav from '@/components/header/VSimNav.vue';
+ import VBrand from '@/components/nav/VBrand.vue';
+ import VHelpAndLogout from '@/components/nav/VHelpAndLogout.vue';
+ import VSimNav from '@/components/nav/VSimNav.vue';
 </script>
