+++ vue/src/components/nav/VSimNav.vue
