+++ vue/src/components/widget/VLongText.vue
@@ -60,4 +60,5 @@
      }
  });
 
+ rawValue.value = field().val;
 </script>
